fixing @internal lib conflcit warning
