+++ packages/flame_tiled/lib/src/renderable_layers/tile_layers/unsupported_layer.dart
@@ -1,0 +1,17 @@
+import 'package:flame_tiled/flame_tiled.dart';
+import 'package:meta/meta.dart';
+
+// Represents a [RenderableLayer] that cannot be parsed by this package.
+@internal
+class UnsupportedLayer extends RenderableLayer {
+  UnsupportedLayer({
+    required super.layer,
+    required super.map,
+    required super.parent,
+    required super.camera,
+    required super.destTileSize,
+  });
+
+  @override
+  void refreshCache() {}
+}
